requirements.txt
aiofiles==22.1.0
aiosqlite==0.20.0
anyio==3.7.1
appnope==0.1.4
arcgis==2.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.3
cycler==0.12.1
dask==2023.3.2
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.54.1
fqdn==1.5.1
fsspec==2024.10.0
geomet==1.1.0
gssapi==1.9.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipykernel==6.29.5
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==7.7.5
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jedi==0.19.2
Jinja2==3.1.2
json5==0.9.28
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client==6.1.12
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-server==1.24.0
jupyter-server-fileid==0.9.3
jupyter-server-ydoc==0.8.0
jupyter-ydoc==0.2.5
jupyterlab==3.6.3
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==1.1.11
keyring==23.13.1
kiwisolver==1.4.7
krb5==0.7.0
locket==1.0.0
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.2
more-itertools==10.5.0
nbclassic==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==6.5.4
notebook-shim==0.2.4
numpy==1.21.6
oauthlib==3.2.2
packaging==24.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.3.6
prometheus-client==0.21.0
prompt-toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
puremagic==1.15
pyarrow==11.0.0
pycparser==2.22
pygments==2.18.0
pylerc==4.0
pyparsing==3.2.0
pyshp==2.3.1
pyspnego==0.8.0
python-certifi-win32==1.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.31.0
requests-gssapi==1.3.0
requests-kerberos==0.15.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.21.0
Send2Trash==1.8.3
setuptools-scm==8.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.4.0
tomli==2.1.0
toolz==1.0.0
tornado==6.4.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
typing-extensions==4.12.2
tzdata==2024.2
ujson==5.7.0
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.1.3
widgetsnbextension==3.6.4
wrapt==1.16.0
y-py==0.6.2
ypy-websocket==0.8.4
zipp==3.21.0
